+++ common/routes/News/containers/NewsContainer.js
@@ -54,6 +54,14 @@ class NewsContainer extends PureComponent {
         }
       }, 100);
     }
+
+    // reload ig embed
+    setTimeout(function () {
+      if (window.instgrm) {
+        window.instgrm.Embeds.process();
+      }
+    }, 100);
+
     this.props.loadInterest();
     window.addEventListener('scroll', this.scrollListener);
   }