+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 而項目則基於 [Semantic Versioning](http://semver.org/lang/zh-TW/).
 
 ## [Unreleased]
+### Changed
+- 調整文件 @esbb48
+- 頁面重整重新載入 IG embed @esbb48
+
 ## 1.0.54 - 2018-01-18
 ### Changed
 - 調整 Header 十週年 Logo @esbb48